+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Current
 
 - Skip unknown fields in mask when applied on a model
 - Added `*` token to fields mask (all remaining fields)
+- Ensure generated endpoints does not collide
 
 0.8.1 (2015-11-27)
 ------------------